refactor: reorganize tests by theme
